fix bug show notification
