+++ .github/workflows/jobspot.yaml
@@ -17,8 +17,8 @@ jobs:
   build:
     runs-on: ubuntu-latest
     env:
-      VERSION_NAME: "1.0.5"
-      VERSION_CODE: 6
+      VERSION_NAME: "1.0.6"
+      VERSION_CODE: 7
 
     steps:
       - name: Checkout code
