+++ android/app/src/main/AndroidManifest.xml
@@ -48,7 +48,7 @@
         </activity>
 
         <meta-data android:name="com.google.firebase.messaging.default_notification_icon"
-            android:resource="@mipmap/ic_launcher" />
+            android:resource="@drawable/ic_launcher" />
 
         <meta-data android:name="com.google.firebase.messaging.default_notification_color"
             android:resource="@color/white" />
