+++ lib/src/presentations/main/screen/main_page.dart
@@ -3,7 +3,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:jobspot/injection.dart';
 import 'package:jobspot/src/core/service/network_service.dart';
-import 'package:jobspot/src/core/service/firebase_messaging_service.dart';
 import 'package:jobspot/src/presentations/main/cubit/main_cubit.dart';
 import 'package:jobspot/src/presentations/main/screen/main_view.dart';
 
@@ -14,9 +13,8 @@ class MainPage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     getIt<NetworkService>().initialise(context);
-    FirebaseMessagingService.setupNotification();
     return BlocProvider(
-      create: (_) => getIt<MainCubit>()..updateToken(),
+      create: (_) => getIt<MainCubit>(),
       child: const MainView(),
     );
   }
