+++ pubspec.yaml
@@ -44,7 +44,7 @@ dependencies:
   intl: ^0.18.0
   get_it: ^7.6.4
   injectable: ^2.3.0
-  fluttertoast: ^8.2.2
+  fluttertoast: ^8.2.3
   flutter_bloc: ^8.1.3
   equatable: ^2.0.5
   firebase_auth: ^4.10.0
